+++ routes/builderRoutes/list.js
@@ -6,6 +6,8 @@ const SeriesReal = require('../../models/builderModels/models-real/Series.js')
 const SeriesOther = require('../../models/builderModels/models-other/Series.js')
 const LessonOther = require('../../models/builderModels/models-other/Lesson.js')
 
+const Topic = require('../../models/nqModels/Topic.js')
+
 var realContent = {
     'rseries': SeriesReal
 }
@@ -31,6 +33,12 @@ module.exports = function (req, res) {
                 })
             } else if (Object.keys(realContent).includes(type)) {
                 realContent[type].find({ users: decodedToken.uid }).exec(function (err, items) {
+                    if (err) console.log(err)
+                    console.log(items)
+                    res.send(items)
+                })
+            } else if (type === 'topic') {
+                Topic.find({}).exec(function (err, items) {
                     if (err) console.log(err)
                     console.log(items)
                     res.send(items)