+++ routes/builderRoutes/search.js
@@ -1,10 +1,13 @@
 const config = require('../../real_config.js')
 const firebase = require('../../firebase.js').builder
+const Fuse = require('fuse.js')
 
 // Import media types
 const SeriesReal = require('../../models/builderModels/models-real/Series.js')
 const SeriesOther = require('../../models/builderModels/models-other/Series.js')
 const LessonOther = require('../../models/builderModels/models-other/Lesson.js')
+
+const Topic = require('../../models/nqModels/Topic.js')
 
 var realContent = {
     'rseries': SeriesReal
@@ -17,6 +20,7 @@ var otherContent = {
 
 module.exports = function (req, res) {
     console.log('type', req.body.type)
+    console.log('search', req.body.terms)
     var token = req.body.token
     
     firebase.verifyID(token)
@@ -55,6 +59,24 @@ module.exports = function (req, res) {
             } else if (Object.keys(realContent).includes(type)) {
                 realContent[type].find({}).where('user', decodedToken.uid).exec(function (err, items) {
                     if (err) console.log(err)
+                    console.log(items)
+                    res.send(items)
+                })
+            } else if (type === 'topic') {
+                console.log('topic search')
+                var query = Topic.find({})
+                if (!options.data) {
+                    query = query.select('_id title')
+                }
+                query.exec(function (err, items) {
+                    if (err) console.log(err)
+                    console.log(items)
+                    // Search with fuse
+                    var options = {
+                        keys: ['title']
+                    }
+                    var search = new Fuse(items, options)
+                    var result = search.search(terms)
                     console.log(items)
                     res.send(items)
                 })